streamlined build process & localized dependencies
you can now press ctrl+b to build after opening the csproj and will be placed in the mod/plugin folder file and dependencies are auto-loaded from the lib folder
