+++ mod/modinfo.json
@@ -1,0 +1,11 @@
+{
+    "id": "pkhead.personalizer",
+    "name": "Personalizer",
+    "version": "1.0",
+    "target_game_version": "v1.9",
+    "authors": "pkhead",
+    "description": "Personalizes your Rain World experience using \"state-of-the-art A.I.\"",
+    "requirements": ["moreslugcats"],
+    "requirements_names": ["More Slugcats Expansion"],
+    "checksum_override_version": false
+}
