+++ plugin/Lizard.cs
@@ -1,0 +1,75 @@
+using BepInEx;
+using UnityEngine;
+using RWCustom;
+
+namespace RWMod
+{
+    public partial class RWMod : BaseUnityPlugin
+    {
+        private void Lizard_ctor(On.Lizard.orig_ctor orig, Lizard self, AbstractCreature creature, World world)
+        {
+            orig(self, creature, world);
+
+            Random.State oldState = Random.state;
+            Random.InitState(creature.ID.RandomSeed);
+
+            // determine if lizard is shiny
+            isShiny[self.abstractCreature.ID] = Random.value < ShinyChance;
+
+            if (isShiny[self.abstractCreature.ID])
+            {
+                // shiny blue lizard, so make the head yellow
+                if (self.Template.type == CreatureTemplate.Type.BlueLizard)
+                {
+                    self.effectColor = Custom.HSL2RGB(
+                        Mathf.Lerp(42f, 59f, Random.value) / 360f,
+                        1f,
+                        Custom.ClampedRandomVariation(0.5f, 0.15f, 0.1f)
+                    );
+                }
+            }
+
+            Random.state = oldState;            
+        }
+
+        private void LizardGraphics_ctor(On.LizardGraphics.orig_ctor orig, LizardGraphics self, PhysicalObject ow)
+        {
+            orig(self, ow);
+
+            // if this is a caramel lizard, then there is a chance that it will be
+            // rendered as black/blue instead
+            if (ModManager.MSC)
+            {
+                if (self.Caramel && IsCreatureShiny(self.lizard.abstractCreature))
+                {
+                    self.lizard.effectColor = new Color(0.0f, 0.0f, 1.0f);
+                }
+            }
+        }
+
+        private void LizardGraphics_ApplyPalette(
+            On.LizardGraphics.orig_ApplyPalette orig,
+            LizardGraphics self,
+            RoomCamera.SpriteLeaser sLeaser,
+            RoomCamera rCam,
+            RoomPalette palette
+        ) {
+            orig(self, sLeaser, rCam, palette);
+
+            if (!self.debugVisualization && IsCreatureShiny(self.lizard.abstractCreature))
+            {
+                // color body of shiny spit lizard (black)
+                if (ModManager.MSC && self.Caramel)
+                {
+                    self.ColorBody(sLeaser, palette.blackColor);
+                }
+
+                // color body of shiny blue lizard (white)
+                else if (self.lizard.Template.type == CreatureTemplate.Type.BlueLizard)
+                {
+                    self.ColorBody(sLeaser, new Color(0.9f, 0.9f, 0.9f));
+                }
+            }
+        }
+    }
+}
